refactor kl divergence expectations
