+++ tests/test_utils.py
@@ -257,7 +257,7 @@ def candidate_test_is_on_temporary_notimplemented_list(context, expectation_type
             # "expect_column_most_common_value_to_be_in_set",
             # "expect_column_sum_to_be_between",
             # "expect_column_min_to_be_between",
-            "expect_column_max_to_be_between",
+            # "expect_column_max_to_be_between",
             "expect_column_chisquare_test_p_value_to_be_greater_than",
             "expect_column_bootstrapped_ks_test_p_value_to_be_greater_than",
             "expect_column_kl_divergence_to_be_less_than",
